Create hashpass.py
If you have changed the OpenSprinkler Pi web interface password from the original default password of "opendoor", the main script needs to have the MD5 Hash of the password replaced with the new password. Place the new password between the double-quote marks on Line 13. Make this file executable and run the program. Enter the results as the new MD5hash in the variables section of ospiLCD-mqtt.py.